+++ include/rmm/mr/device/detail/free_list.hpp
@@ -33,9 +33,38 @@ namespace detail {
  *        of memory allocated from the heap (or upstream allocator).
  */
 struct block {
-  char* ptr;     ///< Raw memory pointer
-  size_t size;   ///< Size in bytes
-  bool is_head;  ///< Indicates whether ptr was allocated from the heap
+  block() = default;
+  block(char* ptr, size_t size, bool is_head) : ptr{ptr}, size_bytes{size}, head{is_head} {}
+
+  /**
+   * @brief Returns the pointer to the memory represented by this block.
+   *
+   * @return the pointer to the memory represented by this block.
+   */
+  inline char* pointer() const { return ptr; }
+
+  /**
+   * @brief Returns the size of the memory represented by this block.
+   *
+   * @return the size in bytes of the memory represented by this block.
+   */
+  inline size_t size() const { return size_bytes; }
+
+  /**
+   * @brief Returns whether this block is the start of an allocation from an upstream allocator.
+   *
+   * A block `b` may not be coalesced with a preceding contiguous block `a` if `b.is_head == true`.
+   *
+   * @return true if this block is the start of an allocation from an upstream allocator.
+   */
+  inline bool is_head() const { return head; }
+
+  /**
+   * @brief Returns whether this block has a non-null pointer.
+   *
+   * @return true if this block has a non-null pointer, false otherwise.
+   */
+  inline bool is_valid() const { return pointer() != nullptr; }
 
   /**
    * @brief Comparison operator to enable comparing blocks and storing in ordered containers.
@@ -46,7 +75,22 @@ struct block {
    * @return true if this block's ptr is < than `rhs` block pointer.
    * @return false if this block's ptr is >= than `rhs` block pointer.
    */
-  bool operator<(block const& rhs) const noexcept { return ptr < rhs.ptr; };
+  inline bool operator<(block const& rhs) const noexcept { return pointer() < rhs.pointer(); };
+
+  /**
+   * @brief Coalesce two contiguous blocks into one.
+   *
+   * `this` must immediately precede `b` and both `this` and `b` must be from the same upstream
+   * allocation. That is, `this->is_contiguous_before(b)`. Otherwise behavior is undefined.
+   *
+   * @param b block to merge
+   * @return block The merged block
+   */
+  inline block merge(block const& b) const noexcept
+  {
+    assert(is_contiguous_before(b));
+    return block(pointer(), size() + b.size(), is_head());
+  }
 
   /**
    * @brief Verifies whether this block can be merged to the beginning of block b.
@@ -55,34 +99,59 @@ struct block {
    * @return true Returns true if this blocks's `ptr` + `size` == `b.ptr`, and `not b.is_head`,
                   false otherwise.
    */
-  bool is_contiguous_before(block const& b) const noexcept
+  inline bool is_contiguous_before(block const& b) const noexcept
   {
-    return (this->ptr + this->size == b.ptr) and not(b.is_head);
+    return (pointer() + size() == b.ptr) and not(b.is_head());
+  }
+
+  /**
+   * @brief Is this block large enough to fit `sz` bytes?
+   *
+   * @param sz The size in bytes to check for fit.
+   * @return true if this block is at least `sz` bytes
+   */
+  inline bool fits(size_t sz) const noexcept { return size() >= sz; }
+
+  /**
+   * @brief Is this block a better fit for `sz` bytes than block `b`?
+   *
+   * @param sz The size in bytes to check for best fit.
+   * @param b The other block to check for fit.
+   * @return true If this block is a tighter fit for `sz` bytes than block `b`.
+   * @return false If this block does not fit `sz` bytes or `b` is a tighter fit.
+   */
+  inline bool is_better_fit(size_t sz, block const& b) const noexcept
+  {
+    return fits(sz) && (size() < b.size() || b.size() < sz);
   }
 
   /**
    * @brief Print this block. For debugging.
    */
-  void print() const { std::cout << reinterpret_cast<void*>(ptr) << " " << size << "B\n"; }
+  void print() const { std::cout << reinterpret_cast<void*>(pointer()) << " " << size() << "B\n"; }
+
+ private:
+  char* ptr{};          ///< Raw memory pointer
+  size_t size_bytes{};  ///< Size in bytes
+  bool head{};          ///< Indicates whether ptr was allocated from the heap
 };
 
 /**
- * @brief Coalesce two contiguous blocks into one.
+ * @brief Comparator for block types based on pointer address.
  *
- * `a` must immediately precede `b` and both `a` and `b` must be from the same upstream allocation.
- * That is, `a.ptr + a.size == b.ptr` and `not b.is_head`. Otherwise behavior is undefined.
- *
- * @param a first block to merge
- * @param b second block to merge
- * @return block The merged block
+ * This comparator allows searching associative containers of blocks by pointer rather than
+ * having to search by the contained type. Saves potentially error-prone temporary construction of
+ * a block when you just want to search by pointer.
  */
-inline block merge_blocks(block const& a, block const& b)
-{
-  assert(a.ptr + a.size == b.ptr);  // non-contiguous blocks
-  assert(not b.is_head);            // Can't merge across upstream allocation boundaries
+template <typename block_type>
+struct compare_blocks {
+  // is_transparent (C++14 feature) allows search key type for set<block_type>::find()
+  using is_transparent = void;
 
-  return block{a.ptr, a.size + b.size};
-}
+  bool operator()(block_type const& lhs, block_type const& rhs) const { return lhs < rhs; }
+  bool operator()(char const* ptr, block_type const& rhs) const { return ptr < rhs.pointer(); }
+  bool operator()(block_type const& lhs, char const* ptr) const { return lhs.pointer() < ptr; };
+};
 
 /**
  * @brief An ordered list of free memory blocks that coalesces contiguous blocks on insertion.
@@ -137,7 +206,7 @@ struct free_list {
     // Find the right place (in ascending ptr order) to insert the block
     // Can't use binary_search because it's a linked list and will be quadratic
     auto const next =
-      std::find_if(blocks.begin(), blocks.end(), [b](block const& i) { return i.ptr > b.ptr; });
+      std::find_if(blocks.begin(), blocks.end(), [b](block const& i) { return b < i; });
     auto const previous = (next == blocks.begin()) ? next : std::prev(next);
 
     // Coalesce with neighboring blocks or insert the new block if it can't be coalesced
@@ -145,13 +214,12 @@ struct free_list {
     bool const merge_next = (next != blocks.end()) && b.is_contiguous_before(*next);
 
     if (merge_prev && merge_next) {
-      *previous = detail::merge_blocks(*previous, b);
-      *previous = detail::merge_blocks(*previous, *next);
+      *previous = previous->merge(b).merge(*next);
       erase(next);
     } else if (merge_prev) {
-      *previous = detail::merge_blocks(*previous, b);
+      *previous = previous->merge(b);
     } else if (merge_next) {
-      *next = detail::merge_blocks(b, *next);
+      *next = b.merge(*next);
     } else {
       insert(next, b);  // cannot be coalesced, just insert
     }
@@ -176,7 +244,7 @@ struct free_list {
    *
    * @param iter An iterator referring to the block to erase.
    */
-  void erase(iterator const& iter) { blocks.erase(iter); }
+  void erase(const_iterator iter) { blocks.erase(iter); }
 
   /**
    * @brief Erase all blocks from the free_list.
@@ -193,11 +261,12 @@ struct free_list {
   block best_fit(size_t size)
   {
     // find best fit block
-    auto const iter = std::min_element(blocks.begin(), blocks.end(), [size](block lhs, block rhs) {
-      return (lhs.size >= size) && ((lhs.size < rhs.size) || (rhs.size < size));
-    });
+    auto const iter =
+      std::min_element(blocks.cbegin(), blocks.cend(), [size](block const& lhs, block const& rhs) {
+        return lhs.is_better_fit(size, rhs);
+      });
 
-    if (iter->size >= size) {
+    if (iter != blocks.end() && iter->fits(size)) {
       // Remove the block from the free_list and return it.
       block const found = *iter;
       erase(iter);